refactor(#741): refactor nameController
