--- api/controllers/v1/NameController.js
@@ -1,12 +1,0 @@
-/**
- */
-
-const nameController = require('../NameController');
-const MappingV1Service = require('../../services/MappingV1Service');
-
-module.exports = {
-  update: (req, res) =>
-    nameController.update(req, res, MappingV1Service.convertToNameModel),
-  setAsMain: (req, res) =>
-    nameController.setAsMain(req, res, MappingV1Service.convertToNameModel),
-};
